Separate data loading from table component
